Update README.md
This commit adds the new online course content (articulate), adds a thank you visual, adds the "what is open data video," and rearranges the older material. The marketing stuff is now higher up and the old PPTs are down below. There are notes on the PPTs explaining what is different from the new version.